Move HTTPConnector to pfio.cache.http
