--- pfio/util/__init__.py
@@ -1,1 +1,0 @@
-from .http import HTTPConnector  # NOQA
